Insect Catch Game
From the start menu, choose to play the game then select a character. The time starts and the user is asked to click on the insects that pop up on the screen. Each insect = 1 point, and insects reappear in a random location and orientation. A message appears after a certain score, however the game runs infinitely.